move App deeper one level
