+++ services/mobile/index.js
@@ -1,5 +1,5 @@
 import registerRootComponent from 'expo/build/launch/registerRootComponent';
 
-import App from './App';
+import { App } from './customer/App';
 
 registerRootComponent(App);
